Replace yapf with my forked version of black.
Yapf was producing inconsistent results when run
on different machines (even with the same config
file and version), so I am replacing it with black.

Mostly I agree with black's opinions on code
formatting, with the notable exception of
indentation level. Black enforces 4 spaces, whereas
I prefer 2. Since the indentation level is not
customizable, I have forked black and changed it to
2 spaces.