Make an empty test suite not fail.
This changes the default behaviour of pytest such that empty test
suites are not considered errors. This is to make it easier to guess
a good shard_count for test targets, as if you over-approximate the
number of shards required, then some of them will be empty and
"fail".

github.com/ChrisCummins/ProGraML/issues/47