Make path expander iterate in sorted order.
This has a performance and memory overhead, though negligible in my
measurements so far. The benefit is that it makes the output of
`format --dry_run` match the user's expectations as with other tools
such as ls and find.

This also enables us to simplify our test code as having a sorted
output order means we don't have to do set comparions.

Signed-off-by: format 2020.01.16 <github.com/ChrisCummins/format>