llvm2graph: Fix node embedding dictionary lookup.
This fixes a serious bug in the node encoder which meant that
dictionary lookups were being performed on the unmodified LLVM text,
e.g. "br label %4", instead of the pre-processed text, e.g. "br label
<%ID>". This causes *all* statement nodes to map to "!UNK" statements.

This is a regression that was introduced in:

    commit 1799757ae9c8f0be00becd70b10133be642a57b2
    Author: Chris Cummins <chrisc.101@gmail.com>
    Date:   Wed Dec 18 13:18:54 2019 +0000

        Work in progress porting graph generation.

        github.com/ChrisCummins/ProGraML/issues/2

The regression was introduced after generating the LLVM datasets that
I have used for experimental work, so already-generated datasets are
not affected. See:

https://github.com/ChrisCummins/phd/commit/1799757ae9c8f0be00becd70b10133be642a57b2#diff-028ee320487c69a9ea788d1f2d071669L321

Signed-off-by: format 2020.01.16 <github.com/ChrisCummins/format>