format: Attempt to resolve proto imports.
The prototocol buffer formatter is sensitive to the directory in which
it is run. This is because it needs to compile the proto file, so must
be able to resolve the paths of "import" statements.

This patch makes the protobuf formatter look for a WORKSPACE file in
the directory chain of a path, and if so, uses that as the root. The
assumption is that proto imports are relative to a bazel workspace
root.

Signed-off-by: format 2020.01.16 <github.com/ChrisCummins/format>