ml4pl: Permit tag names in --restore_from flag.
This extends the CheckpointReference class to contain *either* a
run ID, or the name of a tag. When a tag is used, it is resolved
to one or more run IDs during Logger.Load(), and supports both
loading the checkpoint with the best validation accuracy, or a
model at a specific epoch number, where an epoch number references
only a single run.

Signed-off-by: format 2020.01.21 <github.com/ChrisCummins/format>