ml4pl: Use --tag as progress bar name when available.
This changes the name of the progress bar when running a
train/val/test loop, test-only loop, or cross-validation loop
to the value of --tag, if it is set. E.g.

    Ggnn:200125115919:cc1:  4%|███ | 13/300 [1:43:05<37:55:59, 475.82s/ epoch

would become:

    my_tag:                 4%|███ | 13/300 [1:43:05<37:55:59, 475.82s/ epoch

if --tag=my_tag. The idea here is that --tag is usually set to
some meaningful value (e.g. 'reachability_ggnn_v1'), making it
easier to remember what a long-running job is doing when
looking at the output.

Signed-off-by: format 2020.01.21 <github.com/ChrisCummins/format>