ml4pl/ggnn: Enable data flow step limit durint test.
This replaces the flag --limit_max_data_flow_steps_during_training
with a flag --limit_max_data_flow_steps which applies to all epoch
types. This is to enable running test sets with limited graph step
counts, useful for debugging.

Because limiting data flow steps is something that the user should
be aware, this flips the default value for this flag to false.

Signed-off-by: format 2020.01.21 <github.com/ChrisCummins/format>