ml4pl/llvm2graph: Add support for struct inlining.
This adds support for inlining struct definitions in IR statements, so
that:

    %struct.bar = type { i32 }
    %2 = alloca %struct.bar*, align 8

is inlined to:

    <%ID> = alloca { i32 }*, align 8

To achieve this, the original IR file is required during node encoding
so that we can use inst2vec's existing struct extractor function,
followed by a rewriting pass on the IR statements in the graph.

The struct inlining is done using simple string manipulation and
likely has corner cases and bugs. In the future we should replace this
fragile method with an implementation using the parsed llvm::Module
during llvm2graph graph construction.

github.com/ChrisCummins/ProGraML/issues/57

Signed-off-by: format 2020.01.16 <github.com/ChrisCummins/format>