programl/llvm: Overhaul node text representation.
This changes the text representation used for LLVM nodes as follows:

* Instructions are identified only by their opcode name, e.g.
  "%5 = add %0, 1" becomes "add".
* Variables and constants are identified only by their type, e.g.
  "i32 0" becomes "i32".
* For backwards compatability (and for inst2vec), the full text is
  available as a "full_text" feature.

The goal of this change is to enable the Node.text field to be used as
the key into a randomly-initialized embedding table. By splitting the
representation of instruction and operands into separate nodes, a
smaller vocabulary should achieve higher coverage over real programs.

github.com/ChrisCummins/ProGraML/issues/55

Signed-off-by: format 20.02.05 <github.com/ChrisCummins/format>