programl: Update LLVM version to 10.0.0.
This upgrades the version of LLVM that ProGraML uses from 6.0.0 to
10.0.0. For the most part, this requires little changes to the
code. Only a few APIs have moved around and most things work the
same.

LLVM IR is backwards compatible from 10.0.0 to 6.0.0, so any old IR
datasets can still be used, but the reverse may not be the case -
users will need to update to this version of the code to load new
graphs.

Signed-off-by: format 2020.06.15 <github.com/ChrisCummins/format>